perf: HomeHeroSection Server Component 분리 및 LCP 이미지 최적화
- HomeHeroSection에서 'use client' 제거, 검색 폼을 HomeHeroSearch로 분리
- LCP 이미지에 sizes prop 추가로 반응형 이미지 요청 최적화
- 히어로 이미지 경로를 AVIF 포맷으로 변경
